move ExtensionMessage.js to the articles folder
