+++ src/articles/ExtensionMessage.js
@@ -1,11 +1,11 @@
-import Modal from "./modal_shared/Modal";
-import Header from "./modal_shared/Header";
-import Heading from "./modal_shared/Heading";
-import Main from "./modal_shared/Main";
-import MainImage from "./modal_shared/MainImage";
+import Modal from "../components/modal_shared/Modal";
+import Header from "../components/modal_shared/Header";
+import Heading from "../components/modal_shared/Heading";
+import Main from "../components/modal_shared/Main";
+import MainImage from "../components/modal_shared/MainImage";
 import Footer from "../pages/info_page_shared/Footer";
-import ButtonContainer from "./modal_shared/ButtonContainer";
-import GoToButton from "./modal_shared/GoToButton";
+import ButtonContainer from "../components/modal_shared/ButtonContainer";
+import GoToButton from "../components/modal_shared/GoToButton";
 import { getExtensionInstallationLinks } from "../utils/misc/extensionInstallationLinks";
 import FileDownloadOutlinedIcon from "@mui/icons-material/FileDownloadOutlined";
 import Feature from "../features/Feature";
