+++ src/articles/FindArticles.js
@@ -6,7 +6,7 @@ import SearchField from "./SearchField";
 import * as s from "./FindArticles.sc";
 import LoadingAnimation from "../components/LoadingAnimation";
 import Reminder from "./Reminder";
-import ExtensionMessage from "../components/ExtensionMessage";
+import ExtensionMessage from "./ExtensionMessage";
 import LocalStorage from "../assorted/LocalStorage";
 import {
   runningInChromeDesktop,
